--- src/components/form/buttons/index.ts
@@ -1,2 +1,0 @@
-export * from './form-button'
-export * from './submit'
