--- src/components/form/index.ts
@@ -1,3 +1,0 @@
-export * from './buttons'
-export * from './inputs'
-export * from './form.element'
