+++ src/components/form/inputs/index.ts
@@ -1,4 +1,2 @@
-export * from './multi-select'
 export * from './check'
-export * from './select'
 export * from './text'
