+++ src/components/index.ts
@@ -1,4 +1,2 @@
-export * from 'components/form'
 export * from 'components/non-form'
-export * from 'components/dialog'
 export * from 'components/element'
