+++ src/components/non-form/index.ts
@@ -1,2 +1,1 @@
 export * from './button'
-export * from './table'
