--- src/modules/login/components/index.ts
@@ -1,1 +1,0 @@
-export * from './custom-login-component'
