--- src/modules/login/cookies/index.ts
@@ -1,1 +1,0 @@
-export const cookiesPath = 'src/modules/login/cookies/exampleCookies.json'
