--- src/modules/login/data/routing.data.ts
@@ -1,1 +1,0 @@
-export const loginExampleUrl = 'urlToLogin'
