--- src/modules/login/model/data.model.ts
@@ -1,4 +1,0 @@
-export interface LoginData {
-    readonly userEmail: string
-    readonly userPassword: string
-}
