--- src/modules/login/model/object.model.ts
@@ -1,7 +1,0 @@
-export interface ExampleLoginObject1 {
-    readonly valueObject1: ExampleLoginObject2[]
-}
-export interface ExampleLoginObject2 {
-    readonly valueObject2: string
-    readonly idObject2: number
-}
