--- src/modules/login/pages/index.ts
@@ -1,1 +1,0 @@
-export * from './login'
