--- src/modules/svgomg/mocks/default.json
@@ -1,6 +1,0 @@
-{
-    "field1": 111,
-    "field2": "field2",
-    "field3": "field3",
-    "field4": "field4"
-}
