--- src/modules/svgomg/model/index.ts
@@ -1,2 +1,0 @@
-export * from './object.model'
-export * from './data.model'
