+++ src/svgomg/components/index.ts
