+++ src/svgomg/data/index.ts
