+++ src/svgomg/data/mocksPath.data.ts
