+++ src/svgomg/data/routing.data.ts
