+++ src/svgomg/model/data.model.ts
