+++ src/svgomg/model/index.ts
