+++ src/svgomg/model/object.model.ts
