+++ src/svgomg/pages/index.ts
